# api/openapi.yaml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  description: REST API to be used to create tableside food ordering applications.
  contact:
    name: Course Discord server
    url: https://discord.gg/dPsnz5u9
    email: mark.wentowski@docsgeek.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - url: http://localhost:80/
    description: Local setup
  - url: '{gitpod-server-url}/'
    description: Gitpod setup
    variables:
      gitpod-server-url:
        default: 'https://8080-<workspace-id>.gitpod.io'
        description: 'Dynamic Gitpod URL. Replace <workspace-id> with your Gitpod workspace ID if needed.'
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get a list of users
      description:
        This operation retrieves a list of users based on the specified
        criteria.
      operationId: getAllUsers
      parameters:
        - name: sort
          in: query
          description: |
            Specifies the field by which the results should be sorted. This parameter allows you to define the sorting order based on a particular field in the dataset. If you include a `sort` value, you must include a `order` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: user_name
        - name: order
          in: query
          description: |
            Determines the order in which the results should be sorted. This parameter specifies whether the sorting should be in ascending or descending order. Valid values are `asc` for ascending order and `desc` for descending order. If you include an `order` value, you must include a `sort` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: asc
        - name: fields
          in: query
          description: |
            Use this parameter to specify which resource properties should be included in the response. Provide a list of schema property names, separated by the `|` delimiter. This allows for precise control over the data returned.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: user_name|email|role
        - name: filter
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator~value`.

            **Basic Equality (eq)**
            - For exact matches, use `fieldName.eq~value`.

            **Comparison Operators:**
            - **Less Than (lt)** Use `fieldName.lt~value`.
            - **Less Than or Equal To (lte)** Use `fieldName.lte~value`.
            - **Greater Than (gt):** Use `fieldName.gt~value`.
            - **Greater Than or Equal To (gte):** Use `fieldName.gte~value`.

            Separate filters using the `|` delimter. Dates should be in `YYYY-MM-DD` format.
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: created_at.gt~2024-05-07|user_name.eq~apiuser01
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            type: integer
          example: 10
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_5'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Users
    post:
      tags:
        - Users
      summary: Create a user account
      description:
        This operation creates a new user with the provided information.
        The API server uses `role` to determine the the access rights of the user.
        The `email` field is used for password reset.
      operationId: postUser
      requestBody:
        $ref: '#/components/requestBodies/CreateUserRequest'
      responses:
        '201':
          description:
            Resource created successfully. The response body contains the
            newly created resource.
          headers:
            Location:
              description: Points to the resource endpoint for the newly created resource
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. The request was malformed or invalid.
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
      x-swagger-router-controller: Users
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a user's details
      description:
        This operation retrieves information about a user identified by
        its unique user ID.
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bc9
        - name: fields
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: user_name|role
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Users
    put:
      tags:
        - Users
      summary: Update a user
      description: This operation updates an existing user with the provided data.
      operationId: putUser
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bc9
      requestBody:
        $ref: '#/components/requestBodies/UpdateUserRequest'
      responses:
        '200':
          description:
            Resource updated successfully. The response body contains the
            updated resource.
          headers:
            Location:
              description: Points to the resource endpoint for the newly created resource
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Users
    delete:
      tags:
        - Users
      summary: Delete a user
      description:
        This operation deletes an existing user identified by its unique
        user ID.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bca
      responses:
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Users
...
components:
  schemas:
...
    CreatedAt:
      type: string
      description: Timestamp when the resource was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
...
    Email:
      type: string
      description:
        The email address of the user that can be used for password resets.
        It cannot be used for querying.
      format: email
      example: johndoe@gmail.com
...
    Id:
      type: string
      description: The unique identifier for the resource.
      readOnly: true
      example: 674d0bf5c28b69001f8e03a1
...
    Password:
      type: string
      description:
        "A password must be at least 8 characters long and contain at least\
        \ one uppercase letter, one lowercase letter, one digit, and one special character\
        \ (@, $, !, %, *, ?, &)."
      example: P@ssw0rd!
...
    Problem:
      required:
        - detail
        - title
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
    Role:
      type: string
      description: The role that determines a user's access rights to resources.
      example: manager
      enum:
        - patron
        - expeditor
        - chef
        - manager
...
    UpdatedAt:
      type: string
      description: Timestamp when the order was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        user_name:
          $ref: '#/components/schemas/UserName'
        email:
          $ref: '#/components/schemas/Email'
        password:
          $ref: '#/components/schemas/Password'
        role:
          $ref: '#/components/schemas/Role'
      example:
        password: P@ssw0rd!
        role: manager
        updated_at: 2024-04-22T10:00:00Z
        user_name: john_doe
        created_at: 2024-04-22T10:00:00Z
        id: 674d0bf5c28b69001f8e03a1
        email: johndoe@gmail.com
    UserName:
      type: string
      description: The user name of the user for login purposes. Cannot contain spaces.
      example: john_doe
...
    inline_response_200_5:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total_results:
          type: integer
          example: 20
      example:
        results:
          - password: P@ssw0rd!
            role: manager
            updated_at: 2024-04-22T10:00:00Z
            user_name: john_doe
            created_at: 2024-04-22T10:00:00Z
            id: 674d0bf5c28b69001f8e03a1
            email: johndoe@gmail.com
          - password: P@ssw0rd!
            role: manager
            updated_at: 2024-04-22T10:00:00Z
            user_name: john_doe
            created_at: 2024-04-22T10:00:00Z
            id: 674d0bf5c28b69001f8e03a1
            email: johndoe@gmail.com
        total_results: 20
...
  requestBodies:
...
    CreateUserRequest:
      description: |
        Request body containing information used to update the resource.
      content:
        application/json:
          schema:
            required:
              - email
              - password
              - role
              - user_name
            type: object
            properties:
              email:
                $ref: '#/components/schemas/Email'
              password:
                $ref: '#/components/schemas/Password'
              role:
                $ref: '#/components/schemas/Role'
              user_name:
                $ref: '#/components/schemas/UserName'
      required: true
    UpdateUserRequest:
      description: |
        Request body containing properties to update the resource.
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                $ref: '#/components/schemas/Email'
              password:
                $ref: '#/components/schemas/Password'
              role:
                $ref: '#/components/schemas/Role'
              user_name:
                $ref: '#/components/schemas/UserName'
      required: true
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.
      scheme: basic